Improve error reporting for BB tag parsing
The bbTag2html method now throws a TextParseException instead of just
returning and empty list. This way error messages can be more precise.

In addition to that parse tag arguments at bbParseTag level in order to
be able to find the matching closing tag as it doesn't contain any argument.